Enhanced documentation.
Fixed n2ksender - BST-94 and BST-D0 now correct in format.
Fixed verify_log.py to check length of bst-d0 using 16-bit length.

Known issues with frame reception - overflow of input buffer suspected.

+++ code/cpp/src/protocols/bem/bem_protocol.cpp
@@ -106,7 +106,7 @@ namespace Actisense
 
 			BemResponse response;
 			response.header.bstId = static_cast<BstId>(datagram.bstId);
-			response.header.storeLength = datagram.storeLength;
+			response.header.storeLength = static_cast<uint8_t>(datagram.storeLength);
 			response.checksumValid = true; /* Assumed validated by BDTP layer */
 
 			const auto& data = datagram.data;
